redirect

+++ client/src/TodoItem.jsx
@@ -1,13 +1,32 @@
-const MY_TOKEN =
-  'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywidXNlcm5hbWUiOiJIZWN0b3IiLCJlbWFpbCI6ImhlY3RvckBleGFtcGxlLmNvbSIsImlhdCI6MTc3NDM2OTkzNiwiZXhwIjoxNzc1MjMzOTM2fQ.4SuuHNZjWtohkRD392OKI5yqt0esMe6peBTHsM9_a_w';
-
 function TodoItem({ item, fetchItems }) {
   async function deleteItem() {
+    const token = localStorage.getItem('token');
+    if (!token) {
+      throw new Error('No token found. Please log in.');
+    }
+
     await fetch(`/api/todos/${item.id}`, {
       method: 'DELETE',
       headers: {
-        Authorization: MY_TOKEN,
+        Authorization: token,
       },
+    });
+    fetchItems();
+  }
+
+  async function toggleCompleted() {
+    const token = localStorage.getItem('token');
+    if (!token) {
+      throw new Error('No token found. Please log in.');
+    }
+
+    await fetch(`/api/todos/${item.id}`, {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: token,
+      },
+      body: JSON.stringify({ ...item, completed: !item.completed }),
     });
     fetchItems();
   }
@@ -18,7 +37,13 @@ function TodoItem({ item, fetchItems }) {
       draggable="true"
       data-completed={item.completed ? 'true' : 'false'}
     >
-      <button className="done-button"></button>
+      <button className="done-button" onClick={toggleCompleted}>
+        {item.completed ? (
+          <img src="./images/icon-check.svg" alt="icon check" />
+        ) : (
+          ''
+        )}
+      </button>
       <span>{item.text}</span>
       <button className="delete-button" onClick={deleteItem}>
         <img src="./images/icon-cross.svg" alt="icon cross" />
